[BUG FIXED] Rscript/pathview/rename_pathview_pic_file.py

+++ Rscript/pathview/rename_pathview_pic_file.py
@@ -10,7 +10,7 @@ import pandas as pd
 
 def parse_input():
     parser = argparse.ArgumentParser()
-    parser.add_argument('-p', '--passed-path-file', dest='input_file', default='passed_path.txt',
+    parser.add_argument('-p', '--passed-path-file', dest='input_file', default='/home/colddata/qinqiang/script/Rscript/pathview/passed_path.txt',
                         help='KEGG_ID 和 KEGG_DEF 文件，默认 passed_path.txt')
     parser.add_argument('-i', '--input-dir', dest='input_dir', default='./', help='所有图片目录，默认当前目录')
     parser.add_argument('-o', '--output-dir', dest='output_dir', help='输出目录，默认当前目录', default='./')
@@ -26,10 +26,12 @@ def rename(input_dir, ref_file):
     ko_png_list = [x for x in os.listdir(input_dir) if x.endswith('ko.data.png') and '_' not in x]
     for png in ko_png_list:
         png_ko_name = os.path.basename(png).split('.')[0]
-        replace_name = ref_df[ref_df['KO'] == png_ko_name]['Def'].to_list()[0]
+        replace_name = ref_df[ref_df['KO'] == png_ko_name]['Def'].to_list()
+        if len(replace_name) == 0:
+            continue
         # print(type(replace_name), replace_name)
         # exit(0)
-        replace_name = f'{png_ko_name}_{replace_name}.ko.data.png'
+        replace_name = f'{png_ko_name}_{replace_name[0]}.ko.data.png'
         print(png, png_ko_name, replace_name)
         # exit(0)
         os.rename(os.path.join(input_dir, png), os.path.join(input_dir, replace_name))
